+++ src/common/js/global.js
@@ -1,0 +1,5 @@
+/*
+ * pwix:i18n/src/common/js/global.js
+ */
+
+pwixI18n = {};